# docs/swagger.yaml
basePath: /
definitions:
...
  structs.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
    type: object
  structs.Transaction:
    properties:
      change:
        type: integer
      created_at: {}
      date: {}
      id:
        type: integer
      paid:
        type: integer
      total:
        type: integer
      updated_at: {}
      user_id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Casheex is a simple RESTful API for a casheer system, designed to support
    basic cashier operations in retail or small business environments.
  title: Casheex
  version: "1.0"
paths:
...
  /api/transactions:
    get:
      description: Retrieve a list of transactions filtered by optional date for the
        authenticated user
      parameters:
      - description: Date filter in format yyyy-mm-dd
        in: query
        name: date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structs.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.Response'
      security:
      - BearerAuth: []
      summary: Get user's transaction history
      tags:
      - Transaction
    post:
      consumes:
      - application/json
      description: Process the transaction and store it in the database
      parameters:
      - description: Transaction Data
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/structs.Transaction'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structs.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/structs.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/structs.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.Response'
      security:
      - BearerAuth: []
      summary: Create a new transaction
      tags:
      - Transaction
  /api/transactions/all:
    get:
      description: Retrieve all transactions, optionally filtered by date
      parameters:
      - description: Date filter in format yyyy-mm-dd
        in: query
        name: date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structs.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.Response'
      security:
      - BearerAuth: []
      summary: Get all transactions (admin)
      tags:
      - Transaction
  /api/transactions/profit:
    get:
      description: Retrieve profit summary between start_date and end_date
      parameters:
      - description: Start date in format yyyy-mm-dd
        in: query
        name: start_date
        required: true
        type: string
      - description: End date in format yyyy-mm-dd
        in: query
        name: end_date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structs.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.Response'
      security:
      - BearerAuth: []
      summary: Calculate profit within date range
      tags:
      - Transaction
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"